RAGpython/requirements.txt
weaviate-client>=3.26.7,<4.0.0
openai
streamlit
requests
